场景优化：planar-shadow，增加旋转摄像头操作
