--- assets/auto-test-case/staticScene/planar-shadow.test.ts.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "4.0.23",
-  "importer": "typescript",
-  "imported": true,
-  "uuid": "3a7cbdb8-d5ad-4034-9183-cee3212c0002",
-  "files": [],
-  "subMetas": {},
-  "userData": {}
-}
